各入力欄を子コンポーネント化
